requirements.txt
pillow
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google-auth==2.6.2
gunicorn==20.1.0
requests==2.27.1
SQLAlchemy==1.4.32
typing-extensions==4.1.1
WTForms==3.0.0
python-dotenv
google-analytics-data
wheel
psycopg2-binary
psycopg2